Refactored redux store, services, configuration & test case
