--- src/config/index.ts
@@ -1,5 +1,0 @@
-import config from './env';
-
-export const Config = {
-  API_URL: config.apiUrl,
-};
