+++ src/services/auth/type.ts
@@ -1,0 +1,8 @@
+export interface Credential {
+  user: string;
+  pwd: string;
+}
+
+export interface AccessToken {
+  accessToken: string;
+}
