+++ src/services/users/type.ts
@@ -1,0 +1,4 @@
+export interface User {
+  username: string;
+  _id: string;
+}
