+++ src/services/users/user.service.ts
@@ -1,0 +1,7 @@
+import { EndpointBuilder } from '@reduxjs/toolkit/dist/query/endpointDefinitions';
+import { User } from './type';
+
+export const getUsers = (builder: EndpointBuilder<any, any, any>) =>
+  builder.query<User[], void>({
+    query: () => '/users',
+  });
