+++ src/store/auth/index.ts
@@ -1,0 +1,3 @@
+export * from './authSlice';
+export * from './types';
+export { default as authReducer } from './authSlice';
