+++ src/store/auth/types.ts
@@ -1,0 +1,4 @@
+export interface AuthState {
+  user: object | null;
+  accessToken: string | null;
+}
